Add files via upload
This is for the Brock Makerspace Design Challenge.  
Designed one piece of a six part model of a ball using Fusion 360 or  in any application you would like to use—sculpt it, enhance it, personalize it in some way,  embed a circuit inside, really anything is okay as long as you leave the magnet holes intact.